+++ src/main/java/com/zepben/evolve/database/sqlite/readers/NetworkServiceReader.kt
@@ -11,6 +11,7 @@ package com.zepben.evolve.database.sqlite.readers
 import com.zepben.evolve.database.sqlite.tables.associations.*
 import com.zepben.evolve.database.sqlite.tables.iec61968.assetinfo.TableCableInfo
 import com.zepben.evolve.database.sqlite.tables.iec61968.assetinfo.TableOverheadWireInfo
+import com.zepben.evolve.database.sqlite.tables.iec61968.assetinfo.TablePowerTransformerInfo
 import com.zepben.evolve.database.sqlite.tables.iec61968.assets.TableAssetOwners
 import com.zepben.evolve.database.sqlite.tables.iec61968.assets.TablePoles
 import com.zepben.evolve.database.sqlite.tables.iec61968.assets.TableStreetlights
@@ -47,6 +48,7 @@ class NetworkServiceReader constructor(getStatement: () -> Statement) : BaseServ
 
         status = status and loadEach<TableCableInfo>("cable info", reader::load)
         status = status and loadEach<TableOverheadWireInfo>("overhead wire info", reader::load)
+        status = status and loadEach<TablePowerTransformerInfo>("power transformer info", reader::load)
         status = status and loadEach<TableLocations>("locations", reader::load)
         status = status and loadEach<TableOrganisations>("organisations", reader::load)
         status = status and loadEach<TableAssetOwners>("asset owners", reader::load)