+++ src/test/java/com/zepben/evolve/database/sqlite/DatabaseSqliteTest.kt
@@ -118,6 +118,11 @@ class DatabaseSqliteTest {
     }
 
     @Test
+    fun `test PowerTransformerInfo Schema`() {
+        validateSchema(SchemaTestNetwork.createPowerTransformerInfoTestServices())
+    }
+
+    @Test
     fun testMetadataDataSourceSchema() {
         validateSchema(SchemaTestNetwork.createDataSourceTestServices())
     }